execs: check EXEC_FLAGS_FORCE_CPU_TRANSPARENCY for host compatible binaries
Currently if a target binary is compatible with the host CPU we run it via
calling dynamic loader explicitly even if the user has configured the
target to use CPU transparency.

So now we check if exec policy has EXEC_FLAGS_FORCE_CPU_TRANSPARENCY in
exec_flags and if it is set run the binary under CPU transparency.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Lauri Aarnio <Lauri.Aarnio@iki.fi>