Rename scratchbox2 to ldbox due to fork.
A new project called ldbox is forked from scratchbox2.
Filenames, variable names and utilities renamed to reflect this change:
* sb2 command renamed to lb;
* sb2d command renabed to lbrdbd (because lbd is already used);
* sb2-* commands renamed to lb-*;
* libsb2 library renamed to liblb;
* internal datatypes, functions, variables also renamed.

The reason of fork is that scratchbox2 is unmaintained for a long time.

The reason for renaming is to boost interest to this project,
that was hampered by being wrongly associated with old scratchbox,
that is heavyweight and difficult to use.

Signed-off-by: Oleg Girko <ol@infoserver.lv>