Workaround for simulated root to create files in non-writable directories.
Simulated root should be allowed to create files in directories
no matter whether the directory has write previssions or not.
Some programs (like rpm) rely on this behaviour.

The workaround is to change directory premissions to allow writing
before retrying creating file in case creating file failed with
EPERM or EACCESS error for the first time.
Then directory premissions are reverted back to original value.

Unfortunately, this approach can lead to race conditions,
but this can't be done atomically.

Signed-off-by: Oleg Girko <ol@infoserver.lv>