Always create user-writable directories in sumulated root mode.
Directories created in emulated root mode should always be writable
by the user, because emulated root should have write access no matter
what (virtual) directory modes are.
This is done by using forced permissions of 0700 added to modes
of the newly created directory in case these modes don't have
"w" or "x" bits for the owner.
This is already done in mkdir() gate, but was forgotten in mkdirat() gate.
This change fixes this omission.

Signed-off-by: Oleg Girko <ol@infoserver.lv>